moved basic models to new package
