+++ src/main/java/org/snake/models/Block.java
@@ -1,5 +1,7 @@
 package org.snake.models;
 
+import org.snake.models.basic.Point;
+
 public class Block extends AbstractFigure {
 
     public Block(Point point) {
