+++ src/main/java/org/snake/models/Figure.java
@@ -1,5 +1,7 @@
 package org.snake.models;
 
+import org.snake.models.basic.Point;
+
 public interface Figure {
     boolean isEmptySpace();
     Point getPoint();
