+++ src/main/java/org/snake/models/basic/Range.java
@@ -1,4 +1,4 @@
-package org.snake.models;
+package org.snake.models.basic;
 
 /**
  * Represents a range with min and max.
